requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/davestgermain/super-csv@8ed0daef3e815aef2f9d456c2ca1b1cb7b4e1bdc#egg=super-csv==0.1.0
amqp==1.4.9
anyjson==0.3.3
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
billiard==3.3.0.23
celery==3.1.26.post2
click==7.0                # via code-annotations
code-annotations==0.3.1
coverage==4.5.3           # via pytest-cov
django-celery==3.3.0
django-crum==0.7.3
django-model-utils==3.1.2
edx-opaque-keys==0.4.4
funcsigs==1.0.2           # via pytest
importlib-metadata==0.15  # via pluggy
jinja2==2.10.1            # via code-annotations
kombu==3.0.37
markupsafe==1.1.1         # via jinja2
more-itertools==5.0.0     # via pytest
pathlib2==2.3.3           # via pytest, pytest-django
pbr==5.2.0
pluggy==0.12.0            # via pytest
py==1.8.0                 # via pytest
pymongo==3.8.0
pytest-cov==2.7.1
pytest-django==3.4.8
pytest==4.5.0             # via pytest-cov, pytest-django
python-slugify==3.0.2     # via code-annotations
pytz==2019.1
pyyaml==5.1               # via code-annotations
scandir==1.10.0           # via pathlib2
six==1.12.0
stevedore==1.30.1
text-unidecode==1.2       # via python-slugify
wcwidth==0.1.7            # via pytest
